+++ module/utility/damage.js
@@ -1,6 +1,32 @@
 import { HERO } from "../config.js";
 import { HEROSYS } from "../herosystem6e.js";
 
+
+
+// DAMAGE CLASS (DC)
+//
+// Different dice of damage are not the same – 2d6 of Killing
+// Damage is much more likely to injure a target than a 2d6
+// Normal Damage attack. For comparisons between damage
+// types, Champions uses Damage Classes (“DC”).
+// 
+// An attack’s DC is based on the number of Active Points in
+// it divided by 5. Thus, a Blast 6d6 and an HKA 2d6 (each with
+// 30 Active Points) each have 6 DCs; if added STR increases that
+// HKA to 3d6+1, it counts as 10 DCs; and so on.
+// 
+// For attacks with Advantages, determine the DCs by
+// making a special Active Point calculation that only counts
+// Advantages that directly affect how the victim takes damage.
+// The GM makes the final call on which Advantages this includes,
+// but typically, the following Advantages qualify: Area Of
+// Effect, Armor Piercing, AVAD, Autofire, Charges (Boostable),
+// Constant, Cumulative, Damage Over Time, Does BODY, Does
+// Knockback, Double Knockback, Increased STUN Multiplier,
+// MegaScale in some instances, Penetrating, Sticky, Time Limit,
+// Transdimensional, Trigger, Uncontrolled, Usable As Attack,
+// Variable Advantage, and Variable Special Effects.
+
 export function determineStrengthDamage(item, effectiveStr) {
     if (!item.system.usesStrength && !item.system.usesTk) { return null; }
 
@@ -15,8 +41,8 @@ export function determineStrengthDamage(item, effectiveStr) {
     const strDice = Math.floor(strDamage / 3);
 
     const pip = strDamage % 3
-    
-    let strTag = (strDice > 0 )? strDice + "d6" : ""; 
+
+    let strTag = (strDice > 0) ? strDice + "d6" : "";
 
     switch (pip) {
         case 1:
@@ -57,7 +83,7 @@ export function getNumberOfEachDice(roll) {
     matches.forEach((current) => {
         const [numDice, diceType] = current.split('d');
 
-        if(diceType == '6') {
+        if (diceType == '6') {
             d6Count += parseInt(numDice);
         } else if (diceType == '3') {
             d3Count += parseInt(numDice);
@@ -72,10 +98,10 @@ export function getNumberOfEachDice(roll) {
 
 export function simplifyDamageRoll(damageRoll) {
     const [d6Count, d3Count, constant] = getNumberOfEachDice(damageRoll)
- 
+
     let output = "";
 
-    if (d6Count !== 0) { output = addTerms(output, d6Count.toString()  + "d6"); }
+    if (d6Count !== 0) { output = addTerms(output, d6Count.toString() + "d6"); }
 
     if (d3Count !== 0) { output = addTerms(output, d3Count.toString() + "d3"); }
 
@@ -85,13 +111,86 @@ export function simplifyDamageRoll(damageRoll) {
 }
 
 export function convertToDC(item, formula) {
+
     const [d6Count, d3Count, constant] = getNumberOfEachDice(formula);
 
     if (!item.system.killing) { return d6Count; }
 
-    const pip = (constant > 0)? 1 : 0
+    const pip = (constant > 0) ? 1 : 0
 
     return parseInt(3 * d6Count + 2 * d3Count + pip || 0)
+}
+
+
+// Determine DC soley from item/attack
+export function convertToDcFromItem(item) {
+
+    let actor = item.actor;
+    let dc = 0;
+
+    // Killing Attack
+    if (item.system.killing) {
+        dc += parseInt(item.system.dice) * 3;
+        switch (item.system.extraDice) {
+            case 'pip':
+                dc += 1;
+                break
+            case 'half':
+                dc += 2;
+                break
+        }
+    } else
+
+    // Normal Attack
+    {
+        dc += parseInt(item.system.dice);
+        switch (item.system.extraDice) {
+            case 'pip':
+                dc += 0.2;
+                break
+            case 'half':
+                dc += 0.5;
+                break
+        }
+    }
+
+    // Combat Skill Levels
+    const csl = CombatSkillLevelsForAttack(item)
+    if (csl && csl.dc > 0) {
+
+        // Simple +1 DC for now (checking on discord to found out rules for use AP ratio)
+        dc += csl.dc;
+
+        // Each DC should roughtly be 5 active points
+        // let dcPerAp =  ((dc * 5) / (item.system.activePointsDc || item.system.activePoints)) || 1;
+        // let ratio = (dcPerAp || 5) / 5;  // Typically 1 to 1 radio
+        // dc += (csl.dc * dcPerAp);
+        // console.log(dcPerAp, dc, csl.dc)
+    }
+
+    // Add in STR
+    if (item.system.usesStrength) {
+        let str = actor.system.characteristics.str.value
+        let str5 = Math.floor(str / 5)
+        dc += str5
+    }
+
+    // Add in TK
+    if (item.system.usesTk) {
+
+        let tkItems = actor.items.filter(o => o.system.rules == "TELEKINESIS");
+        let str = 0
+        for (const item of tkItems) {
+            str += parseInt(item.system.LEVELS.value) || 0
+        }
+        let str5 = Math.floor(str / 5)
+        dc += str5
+    }
+
+
+
+
+    return dc;
 }
 
 export function convertFromDC(item, DC) {
@@ -105,7 +204,7 @@ export function convertFromDC(item, DC) {
 
     let output = "";
 
-    if (d6Count !== 0) { output = addTerms(output, d6Count.toString()  + "d6"); }
+    if (d6Count !== 0) { output = addTerms(output, d6Count.toString() + "d6"); }
 
     if (d3Count !== 0) { output = addTerms(output, d3Count.toString() + "d3"); }
 
@@ -119,7 +218,7 @@ export function addTerms(term1, term2) {
         return (term !== "") && (term !== null)
     }
 
-    let output = (isValid(term1))? term1 : "";
+    let output = (isValid(term1)) ? term1 : "";
 
     if (isValid(term1) && isValid(term2)) { output += " + "; }
 
@@ -152,7 +251,7 @@ export async function handleDamageNegation(item, damageResult, options) {
             }
             else if (term instanceof NumericTerm) {
                 switch (nextSign) {
-                    case ("+") : {
+                    case ("+"): {
                         newTotal += term.number
                         break;
                     }
@@ -188,12 +287,12 @@ export async function handleDamageNegation(item, damageResult, options) {
     }
 
     if (options.damageNegationValue >= 3) {
-        damageResult.terms[0].results = damageResult.terms[0].results.slice(0, -Math.floor(options.damageNegationValue /3));
+        damageResult.terms[0].results = damageResult.terms[0].results.slice(0, -Math.floor(options.damageNegationValue / 3));
     }
 
     const remainder = options.damageNegationValue % 3
 
-    switch(remainder) {
+    switch (remainder) {
         case (2): {
             const formula = damageResult.terms[0].results.length - 1 + "d6 + 1"
             return await newDamageRoll(formula, damageResult)
@@ -207,7 +306,7 @@ export async function handleDamageNegation(item, damageResult, options) {
         case (0): {
             return await newDamageRoll(damageResult.terms[0].results.length + "d6", damageResult)
         }
-        
+
         default: {
             console.warn("Uhandled Damage Negation");
             return damageResult;
@@ -223,4 +322,27 @@ export class HeroRoll extends Roll {
     setTotal(newTotal) {
         this._total = newTotal;
     }
+}
+
+export function CombatSkillLevelsForAttack(item) {
+    let result = {
+        ocv: 0,
+        dcv: 0,
+        dmcv: 0,
+        omcv: 0,
+        dc: 0,
+    }
+
+    let csl = item.actor.items.find(o => o.system.XMLID === "COMBAT_LEVELS" && o.system.attacks && o.system.attacks[item.id])
+    if (csl) {
+        for (let i = 0; i < parseInt(csl.system.LEVELS.value); i++) {
+            result[csl.system.csl[i]] = (result[csl.system.csl[i]] || 0) + 1;
+        }
+        result.item = csl;
+    }
+
+    // Takes 2 CLS for +1 DC
+    result.dc = Math.floor(result.dc / 2);
+    
+    return result;
 }